Add support for Rocks of Daggerfall 0.2 mod
This is incompatible with Rocks of Daggerfall: Gavaudon, since they have
models in common that have been modified. Until Gavaudon is
re-implemented, I'll try to keep the support for both, but you'll have
to choose which one to install.

Rocks of Daggerfall 0.2 comes with a bundled version of Location Loader.
I use it only if you didn't choose to also install Location Loader on
its own (in case it got an upgrade, or whatnot)